+++ HISTORY.rst
@@ -3,6 +3,17 @@
 Release History
 ---------------
 
+1.17.12 (2021-03-15)
++++++++++++++++++++
+
+**Improvements**
+
+- Logging control cleanup / docs
+
+**Bug Fixes**
+
+- missing if in process.py to check order status
+
 1.17.11 (2021-03-12)
 +++++++++++++++++++